\name{cor.func}
\alias{cor.func}
\title{
 internal functions from Dr. Tibshirani's software package GSA
}
\description{
 internal functions from Dr. Tibshirani's software package GSA
}
\references{
Bair, Eric, Trevor Hastie, Debashis Paul, and Robert Tibshirani. Prediction by Supervised Principal Components.Journal of the American Statistical Association 101, no. 473 (March 2006): 119-137. doi:10.1198/016214505000000628. 
}