% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Clustering.R
\name{MAPE.Clustering}
\alias{MAPE.Clustering}
\title{Output pathway cluster results}
\usage{
MAPE.Clustering(
  summary,
  Num_Clusters = 3,
  kappa.result = kappa.result,
  Num_of_gene_lists,
  genelist = NULL,
  pathway,
  enrichment,
  method,
  software,
  sil_cut = 0.1,
  n.text.permute = 1000,
  output_dir = getwd(),
  parallel = FALSE
)
}
\arguments{
\item{Num_Clusters}{number of clusters}

\item{sil_cut}{silhouette cutoff to determinie cluster tightness}

\item{n.text.permute}{text mining permutation times}
}
\value{
The pathway cluster results in a csv file.
}
\description{
This is the third major function in the MetaPath2.0 package which outputs the pathway
cluster results with text mining anotation.
}
\examples{
library(MetaDE)
library(MetaPath)
data(Psychiatry_disease_filtered)
data(pathways)
CPI_result = MAPE2.0(arraydata = Psychiatry_diseases$expr, clinical.data = Psychiatry_diseases$clinical,
                    label = "response",pmtx = NULL,pathway = c(Biocarta.genesets,GOBP.genesets,GOCC.genesets,GOMF.genesets,
                    KEGG.genesets,Reactome.genesets),data.type ="continuous", resp.type = "twoclass",method = "CPI",
                    ind.method = rep("limma",length(Psychiatry_diseases$expr)),paired = rep(FALSE,length(Psychiatry_diseases$expr)),
                    select.group = c("CASE","CTRL"),ref.level ="CTRL",tail="abs",
                    enrichment = "Fisher's exact", DEgene.number = 400,stat = "AW Fisher")
library(irr)
library(ConsensusClusterPlus)
set.seed(15213)
CPI.kappa_result = MAPE.Kappa(summary = CPI_result$summary,pathway = CPI_result$pathway,
                              max_k = 15, q_cutoff = 0.0005,software = CPI_result$method)
library(cluster)
library(gplots)
library(shape)
data(hashtb)
set.seed(15213)
MAPE.Clustering(MAPE.Clustering(summary=CPI_result$summary,Num_Clusters = 8,
                                kappa.result = CPI.kappa_result$kappa,sil_cut=0.1,
                                Num_of_gene_lists=CPI_result$Num_of_gene_lists,genelist =CPI_result$genelist,
                                pathway=CPI_result$pathway, enrichment=CPI_result$enrichment,
                                method=CPI.kappa_result$method,software=CPI_result$method,
                                n.text.permute = 10000)
}
