% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Kappa.R
\name{MAPE.Kappa}
\alias{MAPE.Kappa}
\title{Calculate the kappa statistics for DE Pathways}
\usage{
MAPE.Kappa(
  summary,
  max_k = 10,
  pathway,
  software = c("CPI", "MAPE"),
  method = c("MAPE_I", "MAPE_G", "MAPE_P"),
  q_cutoff = 0.1,
  output_dir = getwd()
)
}
\arguments{
\item{method}{The meta qvalue used for qvalue cutoff. It is one of the three values:
'MAPE_I','MAPE_G','MAPE_P'.}

\item{q_cutoff}{All the pathways with qvalue smaller than q_cutoff will be used for
kappa statistics calculation.}
}
\value{
The matrix of kappa statistics between each DE pathways
}
\description{
This is the second major function in the MetaPath2.0 package which calculates the pairwise
cohen's kappa statistics between significant pathway analysis results.
}
\examples{
data(Psychiatry_disease_filtered)
data(pathways)
CPI_result = MAPE2.0(arraydata = Psychiatry_diseases$expr, clinical.data = Psychiatry_diseases$clinical,
                    label = "response",pmtx = NULL,pathway = c(Biocarta.genesets,GOBP.genesets,GOCC.genesets,GOMF.genesets,
                    KEGG.genesets,Reactome.genesets),data.type ="continuous", resp.type = "twoclass",method = "CPI",
                    ind.method = rep("limma",length(Psychiatry_diseases$expr)),paired = rep(FALSE,length(Psychiatry_diseases$expr)),
                    select.group = c("CASE","CTRL"),ref.level ="CTRL",tail="abs",
                    enrichment = "Fisher's exact", DEgene.number = 400,stat = "AW Fisher")
library(irr)
library(ConsensusClusterPlus)
set.seed(15213)
CPI.kappa_result = MAPE.Kappa(summary = CPI_result$summary,pathway = CPI_result$pathway,
                              max_k = 15, q_cutoff = 0.0005,software = CPI_result$method)
}
\author{
Chien-wei Lin and George Tseng.
}
