\name{MAQC}
\alias{MAQC}
\docType{data}
\title{
The data sets from MAQC project.
}
\description{
This is the microarray data sets from MAQC project.
}

\references{
Popovici, Vlad, Weijie Chen, Brandon G Gallas, Christos Hatzis, Weiwei Shi, Frank W Samuelson, Yuri Nikolsky, et al. Effect of training-sample size and classification difficulty on the accuracy of genomic predictors. Breast cancer research : BCR 12, no. 1 (January 2010): R5. doi:10.1186/bcr2468. http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2880423&tool=pmcentrez&rendertype=abstract.
}
\examples{
data(MAQC)
}
\keyword{datasets}
