\name{pathway.DB}
\alias{pathway.DB}
\docType{data}
\title{
   An example of pathway database.
}
\description{
  This data set is an example of gene set database in a GeneSetCollection format defined by GSEABase. 
  This database is the C2 collection of Molecular Signatures Database provided by Broad institute(http://www.broadinstitute.org/gsea). 

}
\usage{data(pathway.DB)}
