% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Clustering.R
\name{MAPE.Clustering}
\alias{MAPE.Clustering}
\title{Output pathway cluster results}
\usage{
MAPE.Clustering(Num_Clusters = 3, kappa.result = kappa.result,
  pathway = all.pathway, enrichment, method, output_dir = "./")
}
\arguments{
\item{Num_Clusters}{number of clusters}
}
\value{
The pathway cluster results in a csv file.
}
\description{
This is the third major function in the MetaPath2.0 package which outputs the pathway
cluster results with text mining anotation.
}

