% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Clustering.R
\name{MAPE.Clustering}
\alias{MAPE.Clustering}
\title{Output pathway cluster results}
\usage{
MAPE.Clustering(summary, Num_Clusters = 3, kappa.result = kappa.result,
  Num_of_gene_lists, genelist = NULL, pathway, enrichment, method, software,
  output_dir = getwd())
}
\arguments{
\item{Num_Clusters}{number of clusters}
}
\value{
The pathway cluster results in a csv file.
}
\description{
This is the third major function in the MetaPath2.0 package which outputs the pathway
cluster results with text mining anotation.
}
\examples{
data(clinical)
data(Leukemia_v2)
data(pathways)
select.group <- c('inv(16)','t(15;17)')
ref.level <- "inv(16)"
data.type <- "continuous"
ind.method <- c('limma','limma','limma')
resp.type <- "twoclass"
MAPE2.0_result_CPI = MAPE2.0(arraydata = Leukemia,clinical.data = clinical,label = "label",
                        resp.type=resp.type,stat='maxP',method = "CPI", enrichment = "Fisher's exact", 
                        DEgene.number = 400,size.min=15,size.max=500,data.type=data.type,
                        ind.method=ind.method,ref.level=ref.level,select.group=select.group)
MAPE.kappa_result = MAPE.Kappa(summary = MAPE2.0_result$summary, software = MAPE2.0_result$method,
                               pathway = MAPE2.0_result$pathway, max_k = 20, q_cutoff = 0.0005,
                               output_dir = tempdir())
MAPE.Clustering(summary = MAPE2.0_result$summary,Num_Clusters=5,
               Num_of_gene_lists = MAPE2.0_result$Num_of_gene_lists,
               genelist = genelist,kappa.result = MAPE.kappa_result$kappa, 
               pathway = MAPE2.0_result$pathway, enrichment = MAPE2.0_result$enrichment,
               method = MAPE.kappa_result$method,software = MAPE2.0_result$method,
               output_dir = getwd())
}

