% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Clustering.R
\name{MAPE.Clustering}
\alias{MAPE.Clustering}
\title{Output pathway cluster results}
\usage{
MAPE.Clustering(summary, Num_Clusters = 3, kappa.result = kappa.result,
  Num_of_gene_lists, genelist = NULL, pathway = all.pathway, enrichment,
  method, software, output_dir = getwd())
}
\arguments{
\item{Num_Clusters}{number of clusters}
}
\value{
The pathway cluster results in a csv file.
}
\description{
This is the third major function in the MetaPath2.0 package which outputs the pathway
cluster results with text mining anotation.
}
\examples{
data(pathways)
data(MetaDEresult)
data(MetaDEresult)
MAPE2.0_result = MAPE2.0(stat='maxP',method = "MAPE", enrichment = "KS", 
                        size.min=15,size.max=500, MetaDE = TRUE,
                        meta.p = meta.res.p$meta.analysis$pval,ind.p = ind.res$p)
MAPE.kappa_result = MAPE.Kappa(summary = MAPE2.0_result$summary, software = MAPE2.0_result$method,
                              pathway = MAPE2.0_result$pathway, max_k = 20, q_cutoff = 0.05,
                              output_dir = tempdir())
genelist = NULL
if (MAPE2.0_result$enrichment == "Fisher's exact" & MAPE2.0_result$method == "CPI"){
 genelist = MAPE2.0_result$genelist
}

MAPE.Clustering(summary = MAPE2.0_result$summary,Num_Clusters=10,
               Num_of_gene_lists = MAPE2.0_result$Num_of_gene_lists,
               genelist = genelist,kappa.result = MAPE.kappa_result$kappa, 
               pathway = MAPE2.0_result$pathway, enrichment = MAPE2.0_result$enrichment,
               method = MAPE.kappa_result$method,software = MAPE2.0_result$method,
               output_dir = tempdir())
}

