% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Kappa.R
\name{MAPE.Kappa}
\alias{MAPE.Kappa}
\title{Calculate the kappa statistics for DE Pathways}
\usage{
MAPE.Kappa(summary, max_k = 20, pathway, software = c("CPI", "MAPE"),
  method = c("MAPE_I", "MAPE_G", "MAPE_P"), q_cutoff = 0.1,
  output_dir = getwd())
}
\arguments{
\item{method}{The meta qvalue used for qvalue cutoff. It is one of the three values:
'MAPE_I','MAPE_G','MAPE_P'.}

\item{q_cutoff}{All the pathways with qvalue smaller than q_cutoff will be used for
kappa statistics calculation.}
}
\value{
The matrix of kappa statistics between each DE pathways
}
\description{
This is the second major function in the MetaPath2.0 package which calculates the pairwise
cohen's kappa statistics between significant pathway analysis results.
}
\examples{
data(clinical)
data(Leukemia_v2)
data(pathways)
select.group <- c('inv(16)','t(15;17)')
ref.level <- "inv(16)"
data.type <- "continuous"
ind.method <- c('limma','limma','limma')
resp.type <- "twoclass"
MAPE2.0_result = MAPE2.0(arraydata = Leukemia,clinical.data = clinical,label = "label",
                        resp.type=resp.type,stat='maxP',method = "CPI", enrichment = "Fisher's exact", 
                        DEgene.number = 400,size.min=15,size.max=500,data.type=data.type,
                        ind.method=ind.method,ref.level=ref.level,select.group=select.group)
MAPE.kappa_result = MAPE.Kappa(summary = MAPE2.0_result$summary, software = MAPE2.0_result$method,
                               pathway = MAPE2.0_result$pathway, max_k = 20, q_cutoff = 0.0005,
                               output_dir = tempdir())
}
\author{
Chien-wei Lin and George Tseng.
}

