% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE_Kappa.R
\name{MAPE.Kappa}
\alias{MAPE.Kappa}
\title{Calculate the kappa statistics for DE Pathways}
\usage{
MAPE.Kappa(summary, max_k = 20, pathway, software = c("CPI", "MAPE"),
  method = c("MAPE_I", "MAPE_G", "MAPE_P"), q_cutoff = 0.1,
  output_dir = getwd())
}
\arguments{
\item{method}{The meta qvalue used for qvalue cutoff. It is one of the three values:
'MAPE_I','MAPE_G','MAPE_P'.}

\item{q_cutoff}{All the pathways with qvalue smaller than q_cutoff will be used for
kappa statistics calculation.}
}
\value{
The matrix of kappa statistics between each DE pathways
}
\description{
This is the second major function in the MetaPath2.0 package which calculates the pairwise
cohen's kappa statistics between significant pathway analysis results.
}
\examples{
data(pathways)
data(MetaDEresult)
data(MetaDEresult)
MAPE2.0_result = MAPE2.0(stat='maxP',method = "MAPE", enrichment = "KS", 
                        size.min=15,size.max=500, MetaDE = TRUE,
                        meta.p = meta.res.p$meta.analysis$pval,ind.p = ind.res$p)
MAPE.kappa_result = MAPE.Kappa(summary = MAPE2.0_result$summary, software = MAPE2.0_result$method,
                               pathway = MAPE2.0_result$pathway, max_k = 20, q_cutoff = 0.05,
                               output_dir = tempdir())
}
\author{
Chien-wei Lin and George Tseng.
}

